+++ cloudinary-core/src/main/java/com/cloudinary/transformation/TextLayerBuilder.java
@@ -1,7 +1,0 @@
-package com.cloudinary.transformation;
-
-/**
- * @deprecated Use {@link TextLayer} instead
- */
-public class TextLayerBuilder extends TextLayer {
-}